backend/requirements.txt
# HARV Backend Requirements
# Python 3.8+

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# AI Model Integrations
openai==1.3.5
anthropic==0.39.0
google-generativeai==0.8.3
python-dotenv==1.0.0

# Authentication
pydantic==2.5.0
pydantic[email]==2.5.0

# Additional utilities
typing-extensions==4.8.0
httpx==0.25.1
anyio==4.0.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Production server (optional)
gunicorn==21.2.0